Revert "Ts build (#216)"
This reverts commit 3be927bf11fdcc3eb76df8e9b4a644e702f3be34.
