+++ test/dyncluster.js
@@ -41,8 +41,8 @@ describe('Dynamic Cluster - Connect URLs', function() {
     });
 
     it('adding cluster performs update', function(done) {
-        var route_port = 25220;
-        var port = 25221;
+        var route_port = 54220;
+        var port = 54221;
 
         // start a new cluster with single server
         servers = nsc.start_cluster([port], route_port, function() {
@@ -66,7 +66,7 @@ describe('Dynamic Cluster - Connect URLs', function() {
                         // give some time for the server to send infos
                         setTimeout(function() {
                             // we should know of 3 servers - the one we connected and the 2 we added
-                            should(nc.servers.length()).be.equal(3);
+                            should(nc.servers.length).be.equal(3);
                             done();
                         }, 1000);
                     });
@@ -98,7 +98,7 @@ describe('Dynamic Cluster - Connect URLs', function() {
             });
             nc.on('connect', function() {
               var found = [];
-              nc.servers.getServers().forEach(function(s) {
+              nc.servers.forEach(function(s) {
                 found.push(s.url.href);
               });
 
@@ -131,7 +131,7 @@ describe('Dynamic Cluster - Connect URLs', function() {
                 });
                 nc.on('connect', function() {
                     var have = [];
-                    nc.servers.getServers().forEach(function(s) {
+                    nc.servers.forEach(function(s) {
                         have.push(s.url.port);
                     });
 
@@ -185,7 +185,7 @@ describe('Dynamic Cluster - Connect URLs', function() {
       });
 
       nc.on('connect', function (c) {
-          c.servers.getServers().should.have.length(10);
+          c.servers.should.have.length(10);
           setTimeout(function() { c.close(); }, 0);
           done();
       });
@@ -199,14 +199,14 @@ describe('Dynamic Cluster - Connect URLs', function() {
 
     servers = nsc.start_cluster(ports, route_port, function() {
       var nc = NATS.connect({
-        url: "nats://127.0.0.1:" + port,
+        uri: "nats://127.0.0.1:" + port,
         reconnectTimeWait: 100,
         servers: ["nats://127.0.0.1:" + (port+1)]
       });
 
       function countImplicit(c) {
         var count = 0;
-        c.servers.getServers().forEach(function(s) {
+        c.servers.forEach(function(s) {
           if(s.implicit) {
             count++;
           }
@@ -216,7 +216,7 @@ describe('Dynamic Cluster - Connect URLs', function() {
 
       nc.on('serversDiscovered', function() {
         if(countImplicit(nc) === 1) {
-          var found = nc.servers.getServers().find(function(s) {
+          var found = nc.servers.find(function(s) {
             return s.url.href === "nats://127.0.0.1:" + (port+3);
           });
           if(found) {
@@ -229,7 +229,7 @@ describe('Dynamic Cluster - Connect URLs', function() {
         if(!testVersion("1.0.7", nc)) {
           done();
         }
-        nc.servers.getServers().should.have.length(3);
+        nc.servers.should.have.length(3);
         countImplicit(nc).should.be.equal(1);
 
         // remove the implicit one
