+++ test/perms.js
@@ -50,7 +50,6 @@ describe('Auth Basics', function() {
             }
         };
         var cf = path.resolve(os.tmpdir(), 'conf-' + nuid.next() + '.conf');
-        console.log(cf);
         fs.writeFile(cf, ncu.j(conf), function(err) {
             if(err) {
                 done(err);
@@ -69,9 +68,8 @@ describe('Auth Basics', function() {
         var nc = NATS.connect({
             port: PORT,
             user: 'bar',
-            pass: 'bar'
+            password: 'bar'
         });
-
 
         var perms = 0;
         nc.on('permission_error', function() {
