+++ test/properties.js
@@ -24,13 +24,13 @@ var NATS = require('../'),
 describe('Base Properties', function() {
 
     it('should have a version property', function() {
-        NATS.VERSION.should.match(/[0-9]+\.[0-9]+\.[0-9]+/);
+        NATS.version.should.match(/[0-9]+\.[0-9]+\.[0-9]+/);
     });
 
     it('should have the same version as package.json', function() {
         // eslint-disable-next-line
         var pkg = require('../package.json');
-        NATS.VERSION.should.equal(pkg.version);
+        NATS.version.should.equal(pkg.version);
     });
 
     it('should have a connect function', function() {
